+++ src/main/java/frc/robot/utils/diag/DiagNavX.java
@@ -15,7 +15,7 @@ import com.kauailabs.navx.frc.AHRS;
 public class DiagNavX extends DiagDistanceTraveled {
     private AHRS navX;
 
-    public DiagNavX(String name, int requiredTravel, AHRS navX) {
+    public DiagNavX(String name, double requiredTravel, AHRS navX) {
         super(name, requiredTravel);
         this.navX = navX;
 
@@ -23,7 +23,7 @@ public class DiagNavX extends DiagDistanceTraveled {
     }
 
     @Override
-    protected int getCurrentValue() {
-        return (int)navX.getAngle();
+    protected double getCurrentValue() {
+        return navX.getAngle();
     }
 }
