+++ src/main/java/frc/robot/utils/diag/DiagTalonSrxEncoder.java
@@ -8,7 +8,7 @@ import com.ctre.phoenix.motorcontrol.can.WPI_TalonSRX;
  * distance from its initial position (measured at initialization or after a reset)
  */
 public class DiagTalonSrxEncoder extends DiagDistanceTraveled {
-    public DiagTalonSrxEncoder(String name, int requiredTravel) {
+    public DiagTalonSrxEncoder(String name, double requiredTravel) {
         super(name, requiredTravel);
     }
 
@@ -21,14 +21,14 @@ public class DiagTalonSrxEncoder extends DiagDistanceTraveled {
     @param requiredTravel  - the required difference between the initial position to qualify for success
     @param talonSRX         - the encoder instance to test
     */
-    public DiagTalonSrxEncoder(String name, int requiredTravel, WPI_TalonSRX talonSRX) {
+    public DiagTalonSrxEncoder(String name, double requiredTravel, WPI_TalonSRX talonSRX) {
         super(name, requiredTravel);
         this.talonSRX = talonSRX;
         reset();
     }
 
     @Override
-    protected int getCurrentValue() {
-        return (int)talonSRX.getSelectedSensorPosition();
+    protected double getCurrentValue() {
+        return talonSRX.getSelectedSensorPosition();
     }
 }
